# docs/swagger.yaml
basePath: /v1
definitions:
  entity.Ad:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      link:
        type: string
      view_count:
        type: integer
    type: object
...
  entity.CreateAdRequest:
    properties:
      id:
        type: string
      image_url:
        type: string
      link:
        type: string
    type: object
...
  v1.response:
    properties:
      error:
        example: message
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact: {}
  description: 'Backend - Nodirbek No''monov     TG: https://t.me/alwaysgolang'
  title: news back-end
  version: "1.0"
paths:
  /ads:
    put:
      description: Edits ad by ID
      parameters:
      - description: Ad details
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/entity.CreateAdRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Update Ad
      tags:
      - ads
  /ads/:
    post:
      consumes:
      - application/json
      description: Create a new ad with the given details
      parameters:
      - description: Ad details
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/entity.CreateAdRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Ad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Create a new ad
      tags:
      - ads
  /ads/{id}:
    delete:
      description: Delete an ad
      parameters:
      - description: ID of the ads to delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Delete an ad
      tags:
      - ads
    get:
      description: returns ads
      parameters:
      - description: ID of the ads to get
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Gets ad details
      tags:
      - ads
  /ads/all:
    get:
      description: Get all ads with view count for admins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Ad'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get all ads
      tags:
      - ads
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"